renamed cdi packages after import
