--- support/cdi/core/src/main/java/org/ops4j/pax/shiro/cdi/ShiroIni.java
@@ -1,52 +1,0 @@
-/*
- * Copyright 2013 Harald Wellmann
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
- * implied.
- *
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.ops4j.pax.shiro.cdi;
-
-import static java.lang.annotation.ElementType.FIELD;
-import static java.lang.annotation.ElementType.METHOD;
-import static java.lang.annotation.ElementType.PARAMETER;
-import static java.lang.annotation.ElementType.TYPE;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import javax.inject.Qualifier;
-
-/**
- * Qualifies managed beans as referenceable from Shiro INI files with a given name.
- * <p>
- * The bean name can be set explicitly with a {@code @Named} qualifier. Otherwise, if
- * {@code Bean.getBeanClass()} is a bean type of the given bean, the bean name will be the simple
- * class name of this class, with the first letter converted to lower case.
- * <p>
- * In all other cases (e.g. for producer methods), Shiro will log a warning and the bean will not be
- * available in the Shiro environment.
- * 
- * @author Harald Wellmann
- * 
- */
-@Qualifier
-@Target({ TYPE, METHOD, PARAMETER, FIELD })
-@Retention(RUNTIME)
-@Documented
-public @interface ShiroIni {
-
-}
