--- support/cdi/core/src/test/java/org/ops4j/pax/shiro/cdi/impl/BeanManagerProviderTest.java
@@ -1,43 +1,0 @@
-/*
- * Copyright 2013 Harald Wellmann
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
- * implied.
- *
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.ops4j.pax.shiro.cdi.impl;
-
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.CoreMatchers.sameInstance;
-import static org.junit.Assert.assertThat;
-
-import javax.enterprise.inject.spi.BeanManager;
-import javax.inject.Inject;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.ops4j.pax.exam.junit.PaxExam;
-
-@RunWith(PaxExam.class)
-public class BeanManagerProviderTest {
-
-    @Inject
-    private BeanManager beanManager;
-
-    @Test
-    public void findBeanManagerWithoutJndi() {
-        BeanManagerProvider.setBeanManager(beanManager);
-        assertThat(BeanManagerProvider.getBeanManager(), is(sameInstance(beanManager)));
-    }
-}
